Improve date filtering in UmbrellaGrid<TItem>
This commit modifies the date filtering logic in the `UmbrellaGrid<TItem>` class. The start date is now set to the beginning of the day (00:00:00) and the end date to the end of the day (23:59:59) by creating new `DateTime` instances. Similar adjustments are made for parsing single date filter values to ensure they represent the end of the day.