Update Obsolete attributes in IApplicationBuilderExtensions
Refactor the `Obsolete` attribute for several methods in the
`IApplicationBuilderExtensions` class to include a second
parameter that triggers compiler errors when used. This
change applies to the methods:
`UseUmbrellaFrontEndCompression`,
`EnsureBlazorDevelopmentPaths`,
`UseUmbrellaApiAcceptLanguage`, and
`UseUmbrellaContentLanguage`.

Additionally, the obsolescence messages have been updated
to provide clearer guidance on alternatives and reasons
for deprecation.